added more corporate scrapers
